Update otte_analysis_log2.Rmd
In course of extended email exchange with authors, they explained the variables I am using for N and to identify cochrane review are not correct. I published an Annotation on their elife paper to ask if they would add the correct variables to allow analysis to be reproduced. They say they cannot as the information is confidential because of Cochrane restrictions.